[RENAME] entity 위치 변
