+++ src/main/java/io/oeid/mogakgo/domain/project/infrastructure/ProjectJpaRepository.java
@@ -1,6 +1,6 @@
 package io.oeid.mogakgo.domain.project.infrastructure;
 
-import io.oeid.mogakgo.domain.project.domain.Project;
+import io.oeid.mogakgo.domain.project.domain.entity.Project;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
